Finished basic layout for chatting screen.
Finished layout of chatting screen and renamed to talk.ejs, I did set a
hard width (in px) on the entry box and buttons, but am currently not
worried about supporting any screen besides my own and my friends so
it's not really a problem. Design for what you need and don't over
engineer.

Started working on login screen.