Created login and bad login files. +css
Updated CSS for login and Bad login to center the form and content on
the page. The index page is going to be a link to register or to login.
Either that or I'll put the forms right there to make it easy, unsure on
this right now.

Once the registration pages are made and the index page, I'll start
working on the C functionality, testing it with Curl and making it all
work together.