Created Register form partial.
Created partial for the register form, going to use the same type of
security as BGI, where there is an admin password to register. It's
not the best, but it's something and I am, as I did with BGI, going to
put it under an authentication check by apache.

I should really install a self signed cert onto my server.