Added registration pages.
Added register and bad register. Also had an idea for how to pull in the
chats easily.

Had an idea to ping users when a new chat comes in. Basic algorithm:

poll server every X seconds for new content.
Server checks last modified time of data file for conversation and
compares to the time the request said it last looked. If the last
modified time is greater, than we return content and it replaces the on
screen page content. If not, we return a 304... I can't remember if
qdecoder supports a 304 or not, so it might end up just being in the
JSON itself as a status: [304|200] type deal.