Fixed Create Conversation function.
Attempting to create a file via fopen + 'r' mode was the problem. I was
stupidly thinking to myself that attempts to open a file for reading
would also create the file. Nope. That makes 0 sense when I thought
about it and fixing said r to an a results in the test successfully
passing.