Tested Conversation Modified function.
The conversation modified takes a time_t and checks to see if the
message has been modified after that or not.

Created internal function to get the conversation filename for 2 users,
need to refactor internals of message.c to use it in more places so
we're not duplicating code everywhere.