Started CGI Scripts.
Updated register form. No password. Need to update login form as well.
It doesn't make sense to have a password as apache/harp are going to
request a password anyway. So really we just have to register a user as
a chat client and continue from there.

created create-user.c (cgi) to register and log a user into a session.