Created polling of conversation.
Polling conversation every 2 seconds or so. Need to make a polling area
to check if the conversation has even been updated or not. If it hasn't
then there is no reason to load the entire conversation.

still. It is a function chat system right now :) -- just not a user
friendly one.