Created History Browser File.
Intent is to use ls and other FileUtils to list out the directory for
the currently logged in user's conversations and then show them in the
unordered list on the history screen. Then whenever a use clicks on the
conversation we'll load that up.

Care will have to be taken so that users only have access to their
conversations.

Either I can have the user select another user from a dropdown to
retrieve their conversations with that user, or I'll have to use str
comparison functions to make sure that it's either ^user-<other> or
<other>-user$. RegEx would come in handy for this type of thing, but
regular expressions aren't really a C thing, moreso a perl one. I wonder
if I should do a perl script for this? --I've never done perl so it
might be interesting, on the other hand it will be a lot easier to just
choose a  dropdown with a username and then use that and the current
sessions user. Makes sense as well from my brain's user point of view of
i want to see the history of myself and user X.